split flags headers and cleanup
