+++ core/include/mmcore/flags/EnumBitMaskOperators.h
@@ -1,0 +1,34 @@
+/**
+* MegaMol
+* Copyright (c) 2021, MegaMol Dev Team
+* All rights reserved.
+*/
+
+#pragma once
+
+#include <type_traits>
+
+// nice idea from here https://wiggling-bits.net/using-enum-classes-as-type-safe-bitmasks/
+
+template<typename Enum>
+struct EnableBitMaskOperators {
+    static const bool enable = false;
+};
+
+template<typename Enum>
+constexpr typename std::enable_if<EnableBitMaskOperators<Enum>::enable, Enum>::type operator|(Enum lhs, Enum rhs) {
+    using underlying = typename std::underlying_type<Enum>::type;
+    return static_cast<Enum>(static_cast<underlying>(lhs) | static_cast<underlying>(rhs));
+}
+
+template<typename Enum>
+constexpr typename std::enable_if<EnableBitMaskOperators<Enum>::enable, Enum>::type operator&(Enum lhs, Enum rhs) {
+    using underlying = typename std::underlying_type<Enum>::type;
+    return static_cast<Enum>(static_cast<underlying>(lhs) & static_cast<underlying>(rhs));
+}
+
+template<typename Enum>
+constexpr typename std::enable_if<EnableBitMaskOperators<Enum>::enable, Enum>::type operator^(Enum lhs, Enum rhs) {
+    using underlying = typename std::underlying_type<Enum>::type;
+    return static_cast<Enum>(static_cast<underlying>(lhs) ^ static_cast<underlying>(rhs));
+}
